[UNTESTED]: deserialize components properly on MessageUpdateEvent
