+++ src/model/channel/message.rs
@@ -22,7 +22,7 @@ use crate::gateway::ShardMessenger;
 #[cfg(feature = "model")]
 use crate::http::{CacheHttp, Http};
 use crate::model::prelude::*;
-use crate::model::utils::{deserialize_val, discord_colours, StrOrInt};
+use crate::model::utils::{deserialize_components, discord_colours, StrOrInt};
 #[cfg(all(feature = "model", feature = "cache"))]
 use crate::utils;
 
@@ -149,54 +149,6 @@ pub struct Message {
     ///
     /// Only present in [`MessageCreateEvent`].
     pub poll: Option<Box<Poll>>,
-}
-
-// Custom deserialize function to deserialize components safely without knocking the whole message
-// out when new components are found but not supported.
-fn deserialize_components<'de, D>(deserializer: D) -> Result<Vec<ActionRow>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct ComponentsVisitor;
-
-    impl<'de> Visitor<'de> for ComponentsVisitor {
-        type Value = Vec<ActionRow>;
-
-        fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            formatter.write_str("a sequence of ActionRow elements")
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-        where
-            A: serde::de::SeqAccess<'de>,
-        {
-            let mut components = Vec::with_capacity(seq.size_hint().unwrap_or_default());
-
-            while let Some(map) = seq.next_element::<JsonMap>()? {
-                // We deserialize only the `kind` field to determine the component type.
-                // We later use this to check if its a supported component before deserializing the
-                // entire payload.
-                let raw_kind =
-                    map.get("type").ok_or_else(|| DeError::missing_field("type"))?.clone();
-                let kind: i64 = deserialize_val(raw_kind)?;
-
-                // Action rows are the only top level component supported in serenity at this time.
-                if kind == 1 {
-                    let value = Value::from(map);
-                    components
-                        .push(ActionRow::deserialize(value).map_err(serde::de::Error::custom)?);
-                } else {
-                    // Top level component is not an action row and cannot be supported on
-                    // serenity@current without breaking changes, so we skip them.
-                    tracing::debug!("Skipping component with unsupported kind: {kind}");
-                }
-            }
-
-            Ok(components)
-        }
-    }
-
-    deserializer.deserialize_seq(ComponentsVisitor)
 }
 
 #[cfg(feature = "model")]
