+++ src/model/event.rs
@@ -15,6 +15,7 @@ use crate::model::utils::{
     deserialize_val,
     emojis,
     members,
+    optional_deserialize_components,
     remove_from_map,
     remove_from_map_opt,
     stickers,
@@ -531,6 +532,7 @@ pub struct MessageUpdateEvent {
     pub interaction_metadata: Option<Option<Box<MessageInteractionMetadata>>>,
     #[serde(default, deserialize_with = "deserialize_some")]
     pub thread: Option<Option<GuildChannel>>,
+    #[serde(deserialize_with = "optional_deserialize_components")]
     pub components: Option<Vec<ActionRow>>,
     pub sticker_items: Option<Vec<StickerItem>>,
     pub position: Option<Option<u64>>,
